+++ src/onegov.core/onegov/core/tests/test_filestorage.py
@@ -67,8 +67,9 @@ def test_filestorage(temporary_directory, redis_url):
         def remember_login(response):
             request.app.remember_identity(response, request, morepath.Identity(
                 userid=request.params.get('userid'),
+                groupid='admins',
                 role='admin',
-                application_id=request.app.application_id
+                application_id=request.app.application_id,
             ))
 
     @App.view(model=Model, name='csrf-token')