+++ src/onegov.core/onegov/core/tests/test_orm.py
@@ -776,6 +776,7 @@ def test_application_retries(postgres_dsn, number_of_retries, redis_url):
         def remember(response):
             identity = morepath.Identity(
                 userid='admin',
+                groupid='admins',
                 role='editor',
                 application_id=request.app.application_id
             )