+++ src/onegov.core/onegov/core/tests/test_request.py
@@ -125,6 +125,7 @@ def test_has_permission(redis_url):
         def remember_identity(response):
             request.app.remember_identity(response, request, morepath.Identity(
                 userid='foo',
+                groupid='admins',
                 role='admin',
                 application_id=request.app.application_id
             ))
@@ -192,8 +193,9 @@ def test_permission_by_view(redis_url):
         def remember_identity(response):
             request.app.remember_identity(response, request, morepath.Identity(
                 userid='foo',
+                groupid='',
                 role='admin',
-                application_id=request.app.application_id
+                application_id=request.app.application_id,
             ))
 
         return 'ok'